refactor:숫자 범위의 역할  lottoNumber로 대체
